refactor(pulse): split App.tsx into focused component modules
