+++ apps/pulse/package.json
@@ -31,6 +31,7 @@
     "@osn/typescript-config": "workspace:*",
     "@solidjs/testing-library": "^0.8.10",
     "@tauri-apps/cli": "^2",
+    "@vitest/coverage-istanbul": "^4.1.0",
     "happy-dom": "^20.8.4",
     "typescript": "~5.6.2",
     "vite": "^6.0.3",
